Update L10_task1 .cpp
Problem : 

Write a function that accepts an array<float, 12> as parameter and return same type array 
with rearranged values as described below. Later, develop and verify its working for a few 
test cases. 
Select the last element in array as a pivot, using logic (sequence of CPP statements, including variable 
assignments, loops and conditions) place it at its position if the array is sorted. Also, array elements 
before it should be less than or equal to it, and array elements after it should be greater than it. The 
elements in the rearranged array are not required to be sorted. The catch is to complete this task with 
simpler logic than sorting logic. You may use several arrays to complete this task. 
parameter array: 23, 54, 12, 19, 67, 9, 45, 32, 90, 58, 19, 48 
one possible return array: 23, 12, 19, 9, 45, 32, 19, 48, 54, 67, 90, 58