Update L10_task4.cpp
 Write a function that accepts an array<float, 20> as its first parameter and return same type 
array as sorted in ascending order. The array received should be sorted in ascending order 
from the beginning up to some index near the middle of array. The rest of the array is also 
sorted in ascending order. The other parameters of the function may be the location from 
which up to which the first array is sorted. Later, develop and verify its working for a few test 
cases. Again, the catch is to complete this task with simpler logic than sorting logic. You may use 
several arrays to complete this task. 
parameter array: 7, 12, 25, 45, 82, 97, 3, 34, 38, 40, 49, 52, 65, 83, 91 
only possible return array: sorted in ascending order 


function should be void and accept array as reference 
and solve the problem without creating any new array.